requirements.txt
async-generator==1.10
attrs==23.1.0
certifi==2022.12.7
cffi==1.15.1
chardet==3.0.4
exceptiongroup==1.1.1
googletrans==3.1.0a0
grpcio-tools==1.51.3
grpcio==1.51.3
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2023.1.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
numpy==1.24.2
outcome==1.2.0
pandas==1.5.3
pip==21.3.1
protobuf==4.22.1
psycopg2==2.9.5
pycparser==2.21
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2022.7.1
rfc3986==1.5.0
selenium==4.8.3
setuptools==60.2.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
ta==0.10.2
trio-websocket==0.10.2
trio==0.22.0
urllib3==1.26.15
wheel==0.37.1
wsproto==1.2.0
